Renamed project and application
* Renamed the project to `config`

    * Since there is only ever one project but many apps within it,
    using a name like `config` is a [commonly used
    convention](https://learndjango.com/tutorials/django-best-practices-projects-vs-apps)
    that can make things less confusing

* Renamed the application to neuronsimulator

   * Naming the application `neuronsimulator` may be a bit general, but
   this is likely a small project that may only have one application

   * Since an application is a python module, the name should follow
   [PEP8
   guidelines](https://www.python.org/dev/peps/pep-0008/#package-and-module-names)
   (it should be short, all-lowercase and not include numbers, dashes,
   periods, spaces, or special characters).

The guidelines used in naming came from from [Django Best Practices:
Projects vs
Apps](https://learndjango.com/tutorials/django-best-practices-projects-vs-apps)
from [LearnDjango](https://learndjango.com/)
