+++ config/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for NeuroSimulator project.
+Django settings for Neuron Simulator project.
 
 Generated by 'django-admin startproject' using Django 3.2.4.
 
@@ -39,13 +39,13 @@ ALLOWED_HOSTS: List[str] = []
 # Application definition
 
 INSTALLED_APPS = [
-    "FuncAtlas.apps.FuncatlasConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    "neuronsimulator",
 ]
 
 MIDDLEWARE = [
@@ -58,7 +58,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "NeuroSimulator.urls"
+ROOT_URLCONF = "config.urls"
 
 TEMPLATES = [
     {
@@ -76,7 +76,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "NeuroSimulator.wsgi.application"
+WSGI_APPLICATION = "config.wsgi.application"
 
 
 # Database
