+++ config/urls.py
@@ -14,11 +14,10 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import include, path
-from FuncAtlas import views
+from django.urls import path
+from neuronsimulator import views
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("", views.home, name="home"),
-    path("FuncAtlas/", include("FuncAtlas.urls"), name="home"),
 ]
