+++ neuronsimulator/admin.py
