+++ neuronsimulator/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class NeuronSimulatorConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "neuronsimulator"
+    verbose_name = "Neuron Simulator"
