+++ neuronsimulator/templates/base.html
@@ -15,7 +15,7 @@
         <link href="{% static 'css/sidebar.css' %}" rel="stylesheet">
         <link href="{% static 'css/extras.css' %}" rel="stylesheet">
 
-        <title>{% block title %}NeuroSimulator{% endblock %}</title>
+        <title>{% block title %}Neuron Simulator{% endblock %}</title>
     </head>
 
     <body>
