+++ neuronsimulator/templates/home.html
@@ -2,7 +2,7 @@
 {% load static %}
 {% block content %}
 <div>
-    <h3>NeuroSimulator Dashboard</h3>
+    <h3>Neuron Simulator Dashboard</h3>
     <br>
     <h5>Note: placeholder; will add contents later</h5>
 </div>
