+++ neuronsimulator/templates/navbar.html
@@ -6,7 +6,7 @@
     <div class="container-fluid">
         <a class="navbar-brand" href="{% url 'home' %}">
             <img src="{% static 'images/logo.png' %}" alt="Icahn Laboratory" width="30" height="24">
-            NeuroSimulator
+            Neuron Simulator
         </a>
 
         <ul class="navbar-nav me-auto">
