+++ neuronsimulator/views.py
